User refactored as AppUser, references edited
