+++ src/main/java/com/example/rentsystem/UsersController.java
@@ -1,6 +1,6 @@
 package com.example.rentsystem;
 
-import com.example.vehicle.VehicleOwnerRepository;
+import com.example.user.VehicleOwnerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
