+++ src/main/java/com/example/vehicle/VehicleController.java
@@ -1,5 +1,6 @@
 package com.example.vehicle;
 
+import com.example.user.VehicleOwnerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
