+++ src/test/java/com/example/user/UserTests.java
@@ -5,16 +5,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 public class UserTests {
     @Autowired
-    private UserController userController;
+    private AppUserController appUserController;
 
     @Test
     public void contextLoads() throws Exception {
-        assertThat(userController).isNotNull();
+        assertThat(appUserController).isNotNull();
     }
 }
